Shift all Contact related commands into contact list command package
