+++ src/main/java/seedu/address/logic/commands/contactlistcommands/DeleteContactCommand.java
@@ -1,19 +1,57 @@
 package seedu.address.logic.commands.contactlistcommands;
 
+import static java.util.Objects.requireNonNull;
+
+import java.util.List;
+
+import seedu.address.commons.core.Messages;
+import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
+import seedu.address.model.contact.Contact;
+
 
 /**
- * Deletes a specific contact from the contact list.
+ * Deletes a contact identified using it's displayed index from the contact list.
  */
 public class DeleteContactCommand extends Command {
+
     public static final String COMMAND_WORD = "deletecontact";
+
+    public static final String MESSAGE_USAGE = COMMAND_WORD
+            + ": Deletes the contact identified by the index number used in the displayed contact list.\n"
+            + "Parameters: INDEX (must be a positive integer)\n"
+            + "Example: " + COMMAND_WORD + " 1";
+
+    public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Contact: %1$s";
+
+    private final Index targetIndex;
+
+    public DeleteContactCommand(Index targetIndex) {
+        this.targetIndex = targetIndex;
+    }
 
     @Override
     public CommandResult execute(Model model) throws CommandException {
-        return null;
+        requireNonNull(model);
+        List<Contact> lastShownList = model.getFilteredContactList();
+
+        if (targetIndex.getZeroBased() >= lastShownList.size()) {
+            throw new CommandException(Messages.MESSAGE_INVALID_CONTACT_DISPLAYED_INDEX);
+        }
+
+        Contact contactToDelete = lastShownList.get(targetIndex.getZeroBased());
+        model.deleteContact(contactToDelete);
+        return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, contactToDelete));
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof DeleteContactCommand // instanceof handles nulls
+                && targetIndex.equals(((DeleteContactCommand) other).targetIndex)); // state check
     }
 
     @Override
