+++ src/main/java/seedu/address/logic/commands/contactlistcommands/EditContactDescriptor.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.contactlistcommands;
 
 import java.util.Optional;
