+++ src/main/java/seedu/address/logic/parser/EditContactParser.java
@@ -7,8 +7,8 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TELEGRAM;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.EditContactCommand;
-import seedu.address.logic.commands.EditContactDescriptor;
+import seedu.address.logic.commands.contactlistcommands.EditContactCommand;
+import seedu.address.logic.commands.contactlistcommands.EditContactDescriptor;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
