+++ src/main/java/seedu/address/logic/parser/contactlistparsers/AddContactParser.java
@@ -1,6 +1,5 @@
 package seedu.address.logic.parser.contactlistparsers;
 
-import seedu.address.logic.commands.contactlistcommands.AddContactCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
