+++ src/main/java/seedu/address/logic/parser/contactlistparsers/DeleteContactParser.java
@@ -1,6 +1,5 @@
 package seedu.address.logic.parser.contactlistparsers;
 
-import seedu.address.logic.commands.contactlistcommands.DeleteContactCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
