+++ src/main/java/seedu/address/logic/parser/contactlistparsers/EditContactParser.java
@@ -1,6 +1,5 @@
 package seedu.address.logic.parser.contactlistparsers;
 
-import seedu.address.logic.commands.contactlistcommands.EditContactCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
